refactor(test_tools): /test/test_tools content moved to ./tools and ./tests/tools
