+++ tests/tools/test_variable_tracker.py
