+++ tools/variable_tracker.py
